+++ init_manager.py
@@ -14,9 +14,7 @@ subprocess.check_call(["systemctl", "start", "docker.service"])
 subprocess.check_call(["systemctl", "enable", "yum-cron"])
 
 # Set values loaded by the template
-store_join_tokens_as_tags = ${store_join_tokens_as_tags}
 s3_bucket = '${s3_bucket}'
-
 instance_index = int('${instance_index}')
 vpc_name = '${vpc_name}'
 
@@ -25,58 +23,7 @@ subprocess.check_call(["hostnamectl", "set-hostname",
                        "manager%d-%s" % (instance_index, vpc_name)])
 
 region_name = json.load(urllib2.urlopen(' http://169.254.169.254/latest/dynamic/instance-identity/document'))['region']
-mac = urllib2.urlopen('http://169.254.169.254/latest/meta-data/mac').read().decode()
-vpc_id = urllib2.urlopen('http://169.254.169.254/latest/meta-data/network/interfaces/macs/%s/vpc-id' % mac).read().decode()
 s3 = boto3.resource('s3', region_name=region_name)
-ec2 = boto3.resource('ec2', region_name=region_name)
-vpc = ec2.vpc(vpc_id)
-
-def store_tokens(manager_token, worker_token):
-    """
-    Stores the tokens in S3 or VPC tag
-    """
-    if store_join_tokens_as_tags:
-        vpc.createTags(
-            Tags = [
-                {
-                    "Key": "swarm_manager_token",
-                    "Value": manager_token
-                },
-                {
-                    "Key": "swarm_worker_token",
-                    "Value": worker_token
-                }
-            ]
-        )
-    else:
-        manager_token_object = s3.Object(s3_bucket, 'manager_token')
-        manager_token_object.put(Body=bytes(manager_token),
-                                StorageClass="ONEZONE_IA")
-
-        worker_token_object = s3.Object(s3_bucket, 'worker_token')
-        worker_token_object.put(Body=bytes(worker_token),
-                                StorageClass="ONEZONE_IA")
-
-
-def get_manager_token():
-    """
-    Gets the manager token from S3 or VPC tag
-    """
-    if store_join_tokens_as_tags:
-    else:
-        manager_token_object = s3.Object(s3_bucket, 'manager_token')
-        manager_token_object.wait_until_exists()
-        return manager_token_object.get()['Body'].read()
-
-def get_worker_token():
-    """
-    Gets the worker token from S3 or VPC tag
-    """
-    if store_join_tokens_as_tags:
-    else:
-        manager_token_object = s3.Object(s3_bucket, 'worker_token')
-        manager_token_object.wait_until_exists()
-        return manager_token_object.get()['Body'].read()
 
 
 def initialize_swarm():
@@ -86,16 +33,22 @@ def initialize_swarm():
     subprocess.check_call(["docker", "swarm", "init"])
     manager_token = subprocess.check_output(
         ["docker", "swarm", "join-token", "-q", "manager"]).strip()
+    manager_token_object = s3.Object(s3_bucket, 'manager_token')
+    manager_token_object.put(Body=bytes(manager_token),
+                             StorageClass="ONEZONE_IA")
+
     worker_token = subprocess.check_output(
         ["docker", "swarm", "join-token", "-q", "worker"]).strip()
-    store_tokens(manager_token, worker_token)
+    worker_token_object = s3.Object(s3_bucket, 'worker_token')
+    worker_token_object.put(Body=bytes(worker_token),
+                            StorageClass="ONEZONE_IA")
 
-if not store_join_tokens_as_tags:
-    try:
-        bucket = s3.Bucket(s3_bucket)
-        bucket.objects.all()
-    except NoCredentialsError as e:
-        time.sleep(5)
+
+try:
+    bucket = s3.Bucket(s3_bucket)
+    bucket.objects.all()
+except NoCredentialsError as e:
+    time.sleep(5)
 
 if instance_index == 0:
     # if this is the first node, check if there exists a manager token and ip1 file
@@ -105,7 +58,8 @@ if instance_index == 0:
     bucket = s3.Bucket(s3_bucket)
     objects = map(lambda o: o.key, bucket.objects.all())
     if 'ip1' in objects and 'manager_token' in objects:
-        manager_token = get_manager_token()
+        manager_token_object = s3.Object(s3_bucket, 'manager_token')
+        manager_token = manager_token_object.get()['Body'].read()
         manager1_ip_object = s3.Object(s3_bucket, 'ip1')
         manager1_ip = manager1_ip_object.get()['Body'].read()
 
@@ -117,10 +71,12 @@ if instance_index == 0:
     else:
         initialize_swarm()
 else:
-    manager_token = get_manager_token()
+    manager_token_object = s3.Object(s3_bucket, 'manager_token')
     manager0_ip_object = s3.Object(s3_bucket, 'ip0')
+    manager_token_object.wait_until_exists()
     manager0_ip_object.wait_until_exists()
 
+    manager_token = manager_token_object.get()['Body'].read()
     manager0_ip = manager0_ip_object.get()['Body'].read()
     subprocess.check_call(
         ["docker", "swarm", "join", "--token", manager_token, manager0_ip])